manifest: move ABI code into file of its own
Make it a little more convenient to work with it.
